refactored api setup
