+++ app/domain/imports/routes.py
@@ -7,7 +7,7 @@ from fastapi import APIRouter, Depends, Path, status
 from pydantic import UUID4
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.core.auth import (
+from app.api.auth import (
     AuthMethod,
     AuthScopes,
     CachingStrategyAuth,
