+++ app/domain/references/routes.py
@@ -8,7 +8,7 @@ from elasticsearch import AsyncElasticsearch
 from fastapi import APIRouter, Depends, Path, Request, status
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.core.auth import (
+from app.api.auth import (
     AuthMethod,
     AuthScopes,
     CachingStrategyAuth,
